racing-league-app/requirements.txt
Flask==2.3.2
Flask-CORS==5.0.0
firebase-admin==6.0.1
python-dotenv==1.0.1
pymongo[srv]==4.6.3
gunicorn==21.2.0
requests==2.31.0